+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 ### 1.1.9
 
 #### Bugfixes
+ * Fix Instant.truncatedTo before 1970 (backport of [threetenbp](https://github.com/ThreeTen/threetenbp/commit/7e55ffc395713f75dc5b5cd9e11451902038c868))
 
 #### dependency updates
  * babel-cli to ^6.11.4